feat: initial 2018 devenv - major version changes
This is the initial release of my awesome wm development environment for 2018.  In this commit there is many new changes including:
- Remixing of another awesomewm config repo (awesome-copycat)
- Theming and prettyness
- Many useful widgets and new keybindings
- Status on systems resources
- Big update to readme with gifs to show the wm in action
- Configuration for xterm

This was a large commit to initalize a functional repo for my personal devenv before moving to a development branch.  This should be a good grounding for further work in 2018. Happy new year!